Changes inside 'BaseActivity.java' and 'NumbersActivity.java'
Three new methods put inside 'BaseActivity.java', which are basically
two setters and one other method called 'putWordsIntoList'.

The setters are just for when we want to set the corresponding values
to our fields when inside the 'NumbersActivity' for example. The other
method is there is so it can extract the words(both 'default' and
translated) and put them into the list with dictionaries.