Add 3 new classes, which help to implement the image fuctionality
Three new classes has been added, in order for the program to be more
easily understood and by thus extended in the future, if the programmer
wants to.

The idea of the new classes is that, instead of us passing a
list of both 'Dictionary' and a separate 'ImageView' in the
'WordAdapter', we can instead pass a list consisting of a
'CustomActivityField', which has both of those things inside of it.